ambeth 2.3

+++ ambeth/Ambeth.Query/ambeth/persistence/IVersionCursor.cs
@@ -9,6 +9,6 @@ namespace De.Osthus.Ambeth.Persistence
 
         new IVersionItem Current { get; }
 
-        sbyte GetAlternateIdCount();
+        int GetAlternateIdCount();
     }
 }
